config: Add capability to disable/hide config options (#594)
This adds two optional properties to config options, `disabled` and
`hidden`, that can access the values of other config options in the
config group and programmatically be hidden or disabled as other config
options change. Essentially, this allows the ability to create "child"
config options that will only appear or only be modifiable if their
"parent" is set a certain way.

I originally had a stronger use case for this with some FRU p4 stuff I
was working on, but I figured out an alternative way to do it without
needing configs. I was going to shelve this, but someone in discord
recently requested an option to have relative (vs. true) directions
called out for P3 Ultimate Relativity, so there is at least a potential
use for this now.

Some completely made up mockups below to illustrate:

Parent Strat is `none`, so both sub-options are hidden:

![demo1](https://github.com/user-attachments/assets/e41733ab-0e21-4552-8ae3-ad0b438a7a01)

Parent strat set; both sub-options visible; but checkbox is false, so
the final sub-option is disabled:

![demo2](https://github.com/user-attachments/assets/2d1510a0-7dfb-410f-a67b-4b21991fc9f5)

Checkbox now true, so the final sub-option is enabled:

![demo3](https://github.com/user-attachments/assets/ed7492d5-fce6-4563-a515-010be922d6dc)