raidboss: update FRU triggerSet comment (#629)
Forgot to update the comment with the last PR.

NB. I have a couple of aoe triggers for p5, but nothing for the big
mechs. I will try to find some time this week to clean up and PR the
stuff I have, but my group was done a while ago and I'm not likely to be
back anytime soon.