plugin: 7.2 struct changes (#630)
Sigs validated and structs otherwise fine. There could be (probably
are?) some changes to the BLM job gauge with the removal of the Astral
Fire/Umbral Ice timer, but I don't play BLM so can't address.

(NB: This includes a small bit of cleanup - I changed the offset for
`CharacterDetails` (and its fields) to match the layout in
aers/FFXIVClientStructs, and `shieldPercentage` was moved into
`CharacterDetails` with the correct offset.)