raidboss/oopsy: Add Suzaku Unreal (#639)
Untested in-game, and I likely won't have time to go back in until
tomorrow, but it should be fine for an initial release. I added the
obvious triggers that were missing, but stuff like the buster swap or
the Interlude towers will take a small amount of logic that I don't want
to attempt without testing.