raidboss: r8s champion's circuit counterclockwise fix (#724)
Based on my logs, this change will fix the current wrong calls when
counterclockwise.

The root of the issue seemed to be that the current trigger iterates
through the array incorrectly when counterclockwise. Clockwise iterates
0 => 4, but counterclock requires different iteration method based on
the current storing of orders.

This PR uses an integer map of the indices to iterate, cheaper operation
than trying to pop and insert.