raidboss/oopsy: Add Seiryu unreal (#759)
This should be self-explanatory. Tested in-game. The old timeline is not
perfect, but fixing it would take a substantial amount of effort for
minimal improvement.

In theory it should be possible to figure out the knockback direction
for the snakes with a `getCombatants` call, but I've already done my
clears for the week. As it stands, I wasn't able to find anything in the
logs that would even be close to useful. The Blue Orochi snake adds
spawn in the center of the instance and their facing is not usable.