raidboss: The Meso Terminal -- timeline and triggers (#782)
Boss 1 is clean and should be mostly fine.

Boss 2 is very best-effort and will rely on the player using their eyes
if they finish early and go to help someone else out. The timeline is
fairly accurate but I'm not sure what all is real and what is due to one
of the bosses being idle at any give point. More research is needed.

Boss 3 was pain. Most of the triggers were reasonably straightforward,
but as you can see from the in-line comments, Bombardment was *very*
complex to figure out initially. I ended up just fixing on the north
safespots, since there are four consistent locations for those.

I don't like the condition mess needed for the first Bombardment, but we
don't have a clean way to trigger for it otherwise. We could maybe count
Memento casts and trigger off the second one with a delay, but I figured
the logic as it stands is a little easier to follow.

Please do offer better suggestions for Bombardment and Keraunography.
I'm not particularly happy with either, but I don't have the math
background to simplify Keraunography, and Bombardment is just really
tricky to work with altogether.

Oopsy will be forthcoming At Some Point.