raidboss: add The Forked Tower: Blood triggers and timelines (#774)
~Adds timelines based on fflogs logs.~ EDIT: Now all network log based.

~I have tested the Demon Tablet in game without the added window 100,50
sync, it could sometimes be off by a few seconds so I added that sync.~

I plan to add more as I get to it... unsure if anyone else has started
these. Might do some basic triggers as well.

Will need to revisit once StatusEffectListForay3 packet captures are
available.

Boss 1 ~~could have an earlier front right/back left call for gravity,
but I couldn't get the data in game in time from what I was trying with
OverlayPlugin. Maybe some other entity or log line has it. Meteor data
seemed unavailable, was expecting some MapEffect at least to say where
they need to go to but even that didn't seem to be present?~~ Completed?

Boss 2 is pretty solid, maybe possible to call where the fireballs will
spawn at, but I didn't look into it much. Could also call a role
positions for that. ~~If debuffs are known we can more thoroughly call
where to go for the oozes.~~

Pronged Passage I am not sure can be made much better. Much of the stuff
is based on kill timings and pull timings.

Boss 3 ~~could use some more triggers for calls on where to go for the
Imitation Blizzards, but I haven't really looked into much of when and
where the data is available. For now those are not called.~~ Completed?

Would be nice to have a timeline for the lockward spawns, perhaps a
counter for each set, this may be feasible with some system log captures
to trigger it and then capturing add combatants, but we don't have a
nice way to run test_timeline on it and the player may periodically be
exiting combat (which may not matter if its not resetting on exit
combat). Could add a simple trigger counting up for the lockward spawns
with suppressSeconds, it may need to do a mod 4 for players to modify
the interrupt callout orders, although I am not certain how this will
work on failed attempts/resets of the authorization...

Boss 4 has a couple tells for the tank positionals, cast ids are used as
one way, but perhaps there could also be a weaponid (glowing weapon) in
the log that could track this? ~~Additionally the boss has a buff for
it, but that seems to not yet be captured in the log.~~ ~~There could be
an earlier call for Rune Axe and Holy Lance if debuffs are known.~~
Right now no call for Holy IV stacks is done, probably not needed but
for new players may be beneficial to tell the stack player where to
stand if they get the debuff. ~~The stack marker coincides with a
lanceblow call, so it would need to be merged if done via the headmarker
and that may be rather noisy. Would also be nice to have a call for the
Assassin's Dagger pattern.~~