ci: set GITHUB_TOKEN env on the octokit client (#830)
fixes: #677

## References

-
https://docs.github.com/en/actions/tutorials/authenticate-with-github_token#permissions-for-the-github_token
-
https://docs.github.com/en/rest/using-the-rest-api/rate-limits-for-the-rest-api?apiVersion=2022-11-28
-
https://pinecone-forest.com/oauth2-v0.0.4/OctoKit-Token-Authentication-Strategy.html

Simply set the `GITHUB_TOKEN` environment variable to increase the API
rate limit from 60 to 5,000 requests per minute.
(The `GITHUB_TOKEN` is a one-time secret token that is automatically
generated by GitHub when workflows are run in GitHub Actions.)

If `GITHUB_TOKEN` is not set in the local environment, this will not
cause any issues.