raidboss: Arcadion 11 normal timeline and triggers (#878)
Tested in-game and against the longest report I could find on FFLogs.

I'm not fully happy about having three separate arrangements for calling
the various weapons, but the tether weapons do their model change
simultaneously, so they explicitly need separate triggers.

We might want to add a new output option for Charybdis-type mechanics.
Elsewhere I've seen it, (where it doesn't have an attached Doom-type
effect,) it's tagged with a bare `aoe` response, but that could
encourage people to waste mitigation on it.